Move _update_scss script to lib/_update
