+++ lib/_update/update-scss/update_scss.js
@@ -1,6 +1,6 @@
 const path = require('path')
 const fs = require('fs')
-const { projectDir, packageDir } = require('../path-utils')
+const { projectDir, packageDir } = require('../../path-utils')
 const appSassPath = path.join(projectDir, 'app', 'assets', 'sass')
 const appSassPatternsPath = path.join(appSassPath, 'patterns')
 const applicationScssPath = path.join(appSassPath, 'application.scss')
