+++ lib/_update/update-scss/update_scss.test.js
@@ -4,7 +4,7 @@ const fs = require('fs')
 const path = require('path')
 const updateKit = require('./update_scss')
 const { appSassPath } = require('./update_scss')
-const { projectDir } = require('../path-utils')
+const { projectDir } = require('../../path-utils')
 
 describe('scripts/update-kit', () => {
   afterEach(() => {
